first steps in registration process
