+++ src/main/java/snowtoslow/work/workProject/configs/RestSecurityConfig.java
@@ -1,4 +1,4 @@
-package snowtoslow.work.workProject.security;
+package snowtoslow.work.workProject.configs;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
