+++ src/main/java/snowtoslow/work/workProject/controller/RegistrationController.java
@@ -1,0 +1,4 @@
+package snowtoslow.work.workProject.controller;
+
+public class RegistrationController {
+}
