+++ src/main/java/snowtoslow/work/workProject/exceptions/NotFountPostException.java
@@ -1,0 +1,11 @@
+package snowtoslow.work.workProject.exceptions.uicontrollers;
+
+public class NotFountPostException extends RuntimeException {
+    public NotFountPostException() {
+        super();
+    }
+
+    public NotFountPostException(String message) {
+        super(message);
+    }
+}
