+++ src/main/java/snowtoslow/work/workProject/rest/CommentController.java
@@ -2,15 +2,20 @@ package snowtoslow.work.workProject.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import snowtoslow.work.workProject.models.Comment;
-import snowtoslow.work.workProject.service.CommentService;
 import snowtoslow.work.workProject.service.CommentServiceImpl;
 
+import java.util.List;
 
-import java.util.List;        //DfSFfFASFASF
 @RestController
-public class CommentController implements CommentService {
+public class CommentController {
 
     @Autowired
     private CommentServiceImpl commentService;
