+++ src/main/java/snowtoslow/work/workProject/service/MyUserDetailsService.java
@@ -1,0 +1,4 @@
+package snowtoslow.work.workProject.service;
+
+public class UserDetailService {
+}
