+++ src/main/java/snowtoslow/work/workProject/transfer/CommentDTO.java
@@ -1,0 +1,4 @@
+package snowtoslow.work.workProject.transfer;
+
+public class CommentDTO {
+}
