+++ src/main/java/snowtoslow/work/workProject/transfer/UserDTO.java
@@ -1,0 +1,4 @@
+package snowtoslow.work.workProject.transfer;
+
+public class UserDTO {
+}
