+++ src/main/resources/templates/login.html
@@ -1,0 +1,26 @@
+<html xmlns:th="https://www.thymeleaf.org">
+
+<title>Please Login</title>
+</head>
+<div th:fragment="content">
+    <form name="f" th:action="@{/login}" method="post">
+        <fieldset>
+            <legend>Please Login</legend>
+            <div th:if="${param.error}" class="alert alert-error">
+                Invalid username and password.
+            </div>
+            <div th:if="${param.logout}" class="alert alert-success">
+                You have been logged out.
+            </div>
+            <label for="username">Username</label>
+            <input type="text" id="username" name="username"/>
+            <label for="password">Password</label>
+            <input type="password" id="password" name="password"/>
+            <div class="form-actions">
+                <button type="submit" class="btn" >Log in</button>
+            </div>
+        </fieldset>
+    </form>
+</div>
+</body>
+</html>
